Service registry.
