+++ .gitignore
@@ -1,6 +1,5 @@
 *\bin
 *\obj
-*\.idea
 *\.vs
 packages/
 *.user
