+++ solutions/.idea/.idea.Waves.Core/.idea/.name
@@ -1,0 +1,1 @@
+Waves.Core
